feat: redirect 조건 변경 (#188)
